requirements.txt
pdfplumber==0.11.7
transformers==4.52.4
openai==1.86.0
python-dotenv==1.1.0
torch==2.7.1
spacy==3.8.7
